mg: fix space_to_tabstop
Since the import of mg in the tree, space_to_tabstop used curbp->w_doto
(the byte offset in the current line) as mean to deduce the current
column for indentation.  This is wrong because it doesn't account for
tab, control characters and octets > 127 (which are all rendered with
more than one column.)  Use instead getcolpos().

ok tb@