support (ignore) universal ctags extended metadata in tagaddress
universal ctags "abuses" the tagaddress, which can be an arbitrary
vi command even if mg assumes it's just a search pattern, to store
metadata after a "vi comment".

While it could be worthwile to support the uctags annotations this
just ignores them so the search patterns work again.

The diff is from Joachim Wiberg "troglobit" mg.

ok tb@